rand_table: use u16 and remove some panic points
This changes some numeric types and operations to make RandState
operations safer. In particular:
 * RAND_TABLE values are all u16
 * index and counter are u16
 * u16->usize does not require any unsafe cast
 * saturating add/sub operations are still correct, and safer

+++ bzlib_private.h
@@ -130,8 +130,8 @@ extern void bz_internal_error ( int errcode );
 
 /* Keep this in sync with bzlib_rust/src/rand_rable.rs:RandState */
 typedef struct {
-   Int32 n_to_go;
-   Int32 table_pos;
+   UInt16 n_to_go;
+   UInt16 table_pos;
 } RandState;
 
 /* Defined in bzlib_rust/src/rand_rable.rs */
